requirements.txt
# Core
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Hardware
pyserial>=3.5

# Hyperliquid
hyperliquid-python-sdk>=0.9.0
eth-account>=0.10.0

# Monitoring & UI
rich>=13.7.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Logging
structlog>=23.2.0

# Data
pandas>=2.0.0
